+++ src/commonMain/kotlin/com/insanusmokrassar/sdi/HelperTools.kt
@@ -4,4 +4,4 @@ import kotlinx.serialization.ImplicitReflectionSerializer
 import kotlinx.serialization.json.Json
 
 @ImplicitReflectionSerializer
-fun loadModule(json: String): Module = Json.nonstrict.parse(Module.serializer(), json)
+fun loadModule(json: String): Module = Json.nonstrict.parse(ModuleDeserializerStrategy, json)