+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## 0.1.2
 
 * All reflection support methods are internal
-* Added `lodModule` functions for more useful working with modules loading
+* Added `loadModule` functions for more useful working with modules loading
+* `Module` now is not serializable and can be created only via `loadModule` functions
 
 ## 0.1.1